Add a move_scorer, but do not integrate it in search yet. Implement logging in unit tests. Preliminary elo estimates:
For 5s+0.1s:

 Score of Loki DEV vs Sargon 1978 1.01: 111 - 53 - 136  [0.597] 300
 ...      Loki DEV playing White: 45 - 28 - 77  [0.557] 150
 ...      Loki DEV playing Black: 66 - 25 - 59  [0.637] 150
 ...      White vs Black: 70 - 94 - 136  [0.460] 300
 Elo difference: 68.0 +/- 29.2, LOS: 100.0 %, DrawRatio: 45.3 %
 SPRT: llr 2.95 (100.3%), lbound -2.94, ubound 2.94 - H1 was accepted

For 10s+0.1s:

 Score of Loki DEV vs Sargon 1978 1.01: 109 - 46 - 100  [0.624] 255
 ...      Loki DEV playing White: 64 - 15 - 48  [0.693] 127
 ...      Loki DEV playing Black: 45 - 31 - 52  [0.555] 128
 ...      White vs Black: 95 - 60 - 100  [0.569] 255
 Elo difference: 87.7 +/- 33.7, LOS: 100.0 %, DrawRatio: 39.2 %
 SPRT: llr 2.96 (100.7%), lbound -2.94, ubound 2.94 - H1 was accepted

Estimated elo of Loki DEV:
 5s+0.1s: 1344 +/- 29
 10s+0.1s: 1364 +/- 34

+++ Loki.Lib/utility/perft.hpp
@@ -45,7 +45,7 @@ namespace loki::utility
 		std::string				m_initial_fen;
 		position::position_t	m_pos;
 		size_t					m_nodes;
-		double					m_nps;
+		int64_t					m_nps;
 
 		size_t perft_test(eDepth d, std::ostream& os);
 		void perft_internal(eDepth d);
